requirements.txt
opencv-python
pillow
numpy
matplotlib
scikit-learn
seaborn
torch
torchvision
plotly
streamlit
tqdm